wip: ike-sa: Use time_t for timestamps
uint32_t is half the size on 64-bit platforms (although unsigned).  While
this should not be a problem on most systems (in particular with monotonic
timestamps that generally start at zero after a reboot), we don't know
what kind of timestamps are actually returned, in particular if
a non-monotonic clock is used (although e.g. with regular Epoch via
CLOCK_REALTIME we should not run into problems until Feb 2106).  And while
it requires 24 bytes more per SA it seems the prudent thing to do.

wip: While the methods and enum are named "*statistic*" we only ever used
this facility for timestamps (scheduled events and last packets), hence the
proposed change to time_t.  But maybe it would be more generic if we changed
the type to uint64_t?