wip: ike: Abort initiation if source IP can't be determined
Without a source IP we'll have problems with NAT-D.

wip: Are there legitimate situations where we can't determine the source
address (limited kernel interfaces perhaps?). I also wonder if the
fallbacks in ike-natd are actually ever used nowadays (probably,
ike_sa_t::resolve_hosts() was way simpler in earlier versions so it was
more common that the source IP was not determined at that point).
So I guess we could just remove all that stuff there.
(actually, 4.1.4 added the source lookup in ike-natd and 4.2.5 added the
same in resolve_hosts - about a year between the two)