libstrongswan: Registration of TUN device plugins
A TUN device plugin is registered by providing its constructor function
to the tun_device_manager. The manager creates instances of the TUN device
using the registered constructor function.

Currently the default POSIX TUN device is not a plugin, but hard-coded
into libstrongswan. If no third party TUN device plugin is registered, then
the built-in POSIX tun device is used.