github: Add AWS-LC CI job
AWS-LC is an OpenSSL derivative which can be used with the openssl plugin. This
PR adds a CI job that resembles the openssl-3 test case. It downloads the
source tarball for an AWS-LC release, builds that source using CMake/Ninja, and
then builds/tests strongSwan using the same technique used by openssl-3.

Related to Issue #1907