openssl: Add thread local cleanup for AWS-LC
The leak-detective CI checks were flagging a memory leak of internal
`RAND_bytes` related state which suggests that the thread's destructor was not
getting called. This adds an explicit call to cleanup the thread specific state
in the openssl plugin's destructor which seems to address the CI failures.

References Issue #1907