wip: key-exchange: Add identifiers for FrodoKEM
wip: we have to wait for (pre-)allocation of identifiers. also, the
current draft doesn't define L1 (the ISO standard apparently does not
define it either) -> if L1 is removed, do the same in the following commits
wip: update the reference in the key_exchange_method_t comment (also for ML-KEM)
wip: the constants should probably be adapted (not use the levels but the
actual lattice dimensions e.g. 976)
wip: the keywords should probably also be changed e.g. to frodokem976 to
match the mlkem* identifiers and the variants maybe at the end
frodokem976a/s (or even frodokem976-aes/shake) to clarify this